refactor(firecracker): split firecracker.rs god object into responsibility modules
firecracker.rs (1423 LOC) -> firecracker/ by responsibility: context
(provider/network/vsock/rootfs singletons + VmAttachments), errors
(VmRuntimeError->SandboxError), types (container status + create/provision
types), provisioning (VmSpec compose: network/vsock/rootfs attach, token,
metadata), reconcile (restart orphan reaping), warm_host (OperatorWarmHost),
warm (engine handle + claim), lifecycle (health/create/start/stop/delete/
status). mod.rs keeps the module doc + shared imports and re-exports the
submodule surface; the pub test hooks (warm_pool_initialized_for_tests,
reconcile_warm_orphans_for_tests) are re-exported at pub under the same
cfg(test|test-utils) gate so integration-test crates still reach them. Pure
code motion, no behavior change; max production module now 293 LOC.
